move project hooks to specific file
