+++ lib/tmuxinator.rb
@@ -10,6 +10,7 @@ require "tmuxinator/wemux_support"
 require "tmuxinator/cli"
 require "tmuxinator/config"
 require "tmuxinator/hooks"
+require "tmuxinator/hooks/project"
 require "tmuxinator/pane"
 require "tmuxinator/project"
 require "tmuxinator/window"
