+++ lib/tmuxinator/hooks/project.rb.new
@@ -1,0 +1,37 @@
+module Tmuxinator
+  module Hooks
+    module Project
+      module_function
+
+      # Commands specified in this hook run when "tmuxinator start project"
+      # command is issued
+      def hook_on_project_start
+        Tmuxinator::Hooks.commands_from yaml["on_project_start"]
+      end
+
+      # Commands specified in this hook run when "tmuxinator start project"
+      # command is issued and there is no tmux session available named "project"
+      def hook_on_project_first_start
+        Tmuxinator::Hooks.commands_from yaml["on_project_first_start"]
+      end
+
+      # Commands specified in this hook run when "tmuxinator start project"
+      # command is issued and there is no tmux session available named "project"
+      def hook_on_project_restart
+        Tmuxinator::Hooks.commands_from yaml["on_project_restart"]
+      end
+
+      # Commands specified in this hook run when you exit from a project ( aka
+      # detach from a tmux session )
+      def hook_on_project_exit
+        Tmuxinator::Hooks.commands_from yaml["on_project_exit"]
+      end
+
+      # Command specified in this hook run when "tmuxinator stop project"
+      # command is issued
+      def hook_on_project_stop
+        Tmuxinator::Hooks.commands_from yaml["on_project_stop"]
+      end
+    end # End Project
+  end # End Hooks
+end # End Tmuxinator
